J'ai pas pu tester j'ai une erreur et je suis fatigué
One or more files do not match the primary file of the database. If you are attempting to attach a database, retry the operation with the correct files.  If this is an existing database, the file may be corrupted and should be restored from a backup.
Could not open new database 'C:\USERS\LEURE\SOURCE\REPOS\AUTOMOBILEMULTISOURCES\AUTOMOBILEMULTISOURCE\APP_DATA\FINALDATABASE.MDF'. CREATE DATABASE is aborted.
An attempt to attach an auto-named database for file C:\Users\leure\source\repos\AutomobileMultiSources\AutomobileMultiSource\App_Data\FinalDatabase.mdf failed. A database with the same name exists, or specified file cannot be opened, or it is located on UNC share.
Log file 'C:\Users\leure\source\repos\AutomobileMultiSources\AutomobileMultiSource\App_Data\FinalDatabase_log.ldf' does not match the primary file.  It may be from a different database or the log may have been rebuilt previously.